requirements.txt
aiofiles==24.1.0
aioice==0.10.1
aiortc==1.13.0
annotated-types==0.7.0
anyio==4.9.0
av==14.4.0
bidict==0.23.1
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
dnspython==2.7.0
fastapi==0.115.12
google-crc32c==1.7.1
h11==0.16.0
httptools==0.6.4
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.2
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0
